[SVELTE] add loading thingy

+++ app/webpack/src/list/RegionsBtn.svelte
@@ -1,22 +1,31 @@
 <script>
+    import {showMessage, withLoading} from "../utils.js";
+
     export let witness;
     export let appLang = "en";
     export let baseUrl = `${window.location.origin}${window.location.pathname}`;
-    // const currentRegionId = parseInt(baseUrl.split('regions/')[1].replace("/", ""));
     export let currentRegionId;
 
     async function deleteRegions() {
-        // todo add confirmation modal
+        const confirmed = await showMessage(
+            appLang === "en" ? "Are you sure you want to delete this record?" : "Voulez-vous vraiment supprimer cet enregistrement?",
+            appLang === "en" ? "Confirm deletion" : "Confirmer la suppression",
+            true
+        );
+
+        if (!confirmed) {
+            return;
+        }
 
         if (typeof currentRegionId !== 'number') {
             throw new Error('Invalid region ID');
         }
         const url = `${window.location.origin}/regions/${currentRegionId}/delete`;
         try {
-            const response = await fetch(url, {
+            const response = await withLoading(() => fetch(url, {
                 method: "DELETE",
                 headers: { "X-CSRFToken": CSRF_TOKEN },
-            });
+            }));
             if (response.status !== 204) {
                 throw new Error(`Failed to delete regions: '${response.statusText}'`);
             }
